// apps/sync-service/src/openapi/order-comment.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * SMILE
 * OpenAPI spec version: 1.0.0
 */
import axios from 'axios';
import type {
  AxiosRequestConfig,
  AxiosResponse
} from 'axios';

export type PostOrderOrderIdCommentBody = {
  comment: string;
};

export type PostOrderOrderIdComment200 = { [key: string]: unknown };

export const getSmile = () => {
/**
 * @summary Memberi komentar
 */
const postOrderOrderIdComment = <TData = AxiosResponse<PostOrderOrderIdComment200>>(
    orderId: string,
    postOrderOrderIdCommentBody: PostOrderOrderIdCommentBody, options?: AxiosRequestConfig
 ): Promise<TData> => {
    return axios.post(
      `/order/${orderId}/comment`,
      postOrderOrderIdCommentBody,options
    );
  }

return {postOrderOrderIdComment}};
export type PostOrderOrderIdCommentResult = AxiosResponse<PostOrderOrderIdComment200>